Triage_Assitant-main/Triage_Assitant-main/requirements.txt
Flask>=3.0.0
mysql-connector-python>=8.2.0
pdfplumber>=0.10.3
requests>=2.31.0
python-dotenv>=1.0.0
Werkzeug>=3.0.1
gunicorn>=21.2.0